learnia-backend/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.12.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8
redis==5.0.1
Pillow==10.1.0

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# Requêtes HTTP
httpx==0.25.2
aiohttp==3.9.1

# IA et traitement (optionnel)
openai==1.3.7
pillow==10.1.0

# Utilitaires
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Logs et monitoring
loguru==0.7.2

# CORS et middleware
fastapi-cors==0.0.6 